statistics: adjust codes.
